Fix a few unnecessary type conversions (#339)
This PR gets rid of a bunch of unneeded type conversions. I was checking out the code and
saw a few of these myself, but most of them were fixed using unconvert.

+++ zip/zip_test.go
@@ -597,7 +597,7 @@ func testZip64(t testing.TB, size int64) *rleBuffer {
 	}
 
 	// read back zip file and check that we get to the end of it
-	r, err := NewReader(buf, int64(buf.Size()))
+	r, err := NewReader(buf, buf.Size())
 	if err != nil {
 		t.Fatal("reader:", err)
 	}
